refactor: refactor rootSaga
