+++ src/store/sagas/root-saga.ts
@@ -1,9 +1,10 @@
-import { fork, all } from 'redux-saga/effects';
+import { all } from 'redux-saga/effects';
+
 import productsSaga from '../products/products.sagas';
 import cartSaga from '../cart/cart.sagas';
 
 export function* rootSaga() {
-  yield all([fork(productsSaga), fork(cartSaga)]); // TODO: jo igy ez?
+  yield all([productsSaga(), cartSaga()]);
 }
 
 export default rootSaga;
